test(api): Move pay tests, add mock for new client call
